+++ gdsfactory/samples/16_component_sequence2.py
@@ -22,8 +22,8 @@ def test_cutback_phase(
 
     # Define a map between symbols and (component, input port, output port)
     symbol_to_component = {
-        "I": (taper, "1", "wg_2"),
-        "O": (taper, "wg_2", "1"),
+        "I": (taper, "W0", "E0"),
+        "O": (taper, "E0", "W0"),
         "S": (wg_short, "W0", "E0"),
         "P": (pm_wg, "W0", "E0"),
         "A": (bend180, "W0", "W1"),